feat: move createNew and sort to navbar
