+++ apps/meteor/client/NavBarV2/NavBarPagesToolbar/hooks/useViewModeItems.spec.tsx
