+++ apps/meteor/client/NavBarV2/NavBarPagesToolbar/index.ts
@@ -2,3 +2,5 @@ export { default as NavBarItemAuditMenu } from './NavBarItemAuditMenu';
 export { default as NavBarItemHomePage } from './NavBarItemHomePage';
 export { default as NavBarItemMarketPlaceMenu } from './NavBarItemMarketPlaceMenu';
 export { default as NavBarItemDirectoryPage } from './NavBarItemDirectoryPage';
+export { default as NavBarItemCreateNew } from './NavBarItemCreateNew';
+export { default as NavBarItemSort } from './NavBarItemSort';
